+++ pillow_heif/reader.py
@@ -113,8 +113,8 @@ class HeifFile:
 
     def thumbnails_all(self, one_for_image=False) -> Iterator[Union[UndecodedHeifThumbnail, HeifThumbnail]]:
         for i in self:
-            for t in i.thumbnails:
-                yield t
+            for thumb in i.thumbnails:
+                yield thumb
                 if one_for_image:
                     break
 
@@ -182,16 +182,25 @@ def is_supported(fp) -> bool:
 
 
 def get_file_mimetype(fp) -> str:
+    """
+    Wrapper around `libheif.get_file_mimetype`.
+
+    :param fp: A filename (string), pathlib.Path object, file object or bytes.
+       The file object must implement ``file.read``, ``file.seek`` and ``file.tell`` methods,
+       and be opened in binary mode.
+    :returns: string with `image/*`. If the format could not be detected, an empty string is returned.
+    """
     __data = _get_bytes(fp, 50)
     return ffi.string(lib.heif_get_file_mime_type(__data, len(__data))).decode()
 
 
 def open_heif(fp, *, apply_transformations: bool = True, convert_hdr_to_8bit: bool = True) -> UndecodedHeifFile:
-    d = _get_bytes(fp)
+    file_data = _get_bytes(fp)
     ctx = lib.heif_context_alloc()
-    collect = _keep_refs(lib.heif_context_free, data=d)
-    ctx = ffi.gc(ctx, collect, size=len(d))
-    return _read_heif_context(ctx, d, apply_transformations, convert_hdr_to_8bit)
+    print(f"CTX CONSTRUCTOR:{ctx}")
+    collect = _keep_refs(lib.heif_context_free, data=file_data)
+    ctx = ffi.gc(ctx, collect, size=len(file_data))
+    return _read_heif_context(ctx, file_data, apply_transformations, convert_hdr_to_8bit)
 
 
 def read_heif(fp, *, apply_transformations: bool = True, convert_hdr_to_8bit: bool = True) -> HeifFile:
@@ -203,16 +212,16 @@ def read_heif(fp, *, apply_transformations: bool = True, convert_hdr_to_8bit: bo
     return heif_file.load()
 
 
-def _get_bytes(fp, length=None):
+def _get_bytes(fp, length=None) -> bytes:
     if isinstance(fp, (str, pathlib.Path)):
-        with builtins.open(fp, "rb") as f:
-            return f.read(length or -1)
+        with builtins.open(fp, "rb") as file:
+            return file.read(length or -1)
     if hasattr(fp, "read"):
         offset = fp.tell() if hasattr(fp, "tell") else None
-        b = fp.read(length or -1)
+        result = fp.read(length or -1)
         if offset is not None and hasattr(fp, "seek"):
             fp.seek(offset)
-        return b
+        return result
     return bytes(fp)[:length]
 
 
@@ -223,24 +232,26 @@ def _keep_refs(destructor, **refs):
     """
 
     def inner(cdata):
+        print(f"DESTRUCTOR:{cdata}")
         return destructor(cdata)
 
     inner._refs = refs
     return inner
 
 
-def _read_heif_context(ctx, d, transforms: bool, to_8bit: bool) -> UndecodedHeifFile:
-    brand = lib.heif_main_brand(d[:12], 12)
-    error = lib.heif_context_read_from_memory_without_copy(ctx, d, len(d), ffi.NULL)
+def _read_heif_context(ctx, data, transforms: bool, to_8bit: bool) -> UndecodedHeifFile:
+    brand = lib.heif_main_brand(data[:12], 12)
+    error = lib.heif_context_read_from_memory_without_copy(ctx, data, len(data), ffi.NULL)
     check_libheif_error(error)
     p_main_id = ffi.new("heif_item_id *")
     error = lib.heif_context_get_primary_image_ID(ctx, p_main_id)
     check_libheif_error(error)
     p_main_handle = ffi.new("struct heif_image_handle **")
     error = lib.heif_context_get_primary_image_handle(ctx, p_main_handle)
+    main_handle = p_main_handle[0]
     check_libheif_error(error)
     collect = _keep_refs(lib.heif_image_handle_release, ctx=ctx)
-    handle = ffi.gc(p_main_handle[0], collect)
+    handle = ffi.gc(main_handle, collect)
     return _read_heif_handle(ctx, p_main_id[0], handle, transforms, to_8bit, brand=brand)
 
 
@@ -296,13 +307,13 @@ def _read_metadata(handle) -> list:
 def _retrieve_exif(metadata: list):
     _result = None
     _purge = []
-    for i, v in enumerate(metadata):
-        if v["type"] == "Exif":
+    for i, md_block in enumerate(metadata):
+        if md_block["type"] == "Exif":
             _purge.append(i)
-            if not _result and v["data"] and v["data"][0:4] == b"Exif":
-                _result = v["data"]
-    for e in reversed(_purge):
-        del metadata[e]
+            if not _result and md_block["data"] and md_block["data"][0:4] == b"Exif":
+                _result = md_block["data"]
+    for i in reversed(_purge):
+        del metadata[i]
     return _result
 
 
@@ -348,6 +359,7 @@ def _read_heif_image(handle, heif_class: Union[UndecodedHeifFile, UndecodedHeifT
     p_options.ignore_transformations = int(not heif_class.transforms)
     p_options.convert_hdr_to_8bit = int(heif_class.to_8bit)
     p_img = ffi.new("struct heif_image **")
+    print(f"CONSTRUCTOR:{handle}")
     error = lib.heif_decode_image(handle, p_img, colorspace, chroma, p_options)
     check_libheif_error(error)
     img = p_img[0]
@@ -371,11 +383,11 @@ def _read_thumbnails(handle, transforms: bool, to_8bit: bool) -> list:
     result: List[Union[UndecodedHeifThumbnail, HeifThumbnail]] = []
     if not options().thumbnails:
         return result
-    n = lib.heif_image_handle_get_number_of_thumbnails(handle)
-    if n == 0:
+    thumbs_count = lib.heif_image_handle_get_number_of_thumbnails(handle)
+    if thumbs_count == 0:
         return result
-    thumbnails_ids = ffi.new("heif_item_id[]", n)
-    lib.heif_image_handle_get_list_of_thumbnail_IDs(handle, thumbnails_ids, n)
+    thumbnails_ids = ffi.new("heif_item_id[]", thumbs_count)
+    lib.heif_image_handle_get_list_of_thumbnail_IDs(handle, thumbnails_ids, thumbs_count)
     for thumbnail_id in thumbnails_ids:
         p_handle = ffi.new("struct heif_image_handle **")
         error = lib.heif_image_handle_get_thumbnail(handle, thumbnail_id, p_handle)
@@ -403,11 +415,11 @@ def _read_thumbnail_handle(handle, transforms: bool, to_8bit: bool, **kwargs) ->
 
 def _get_other_top_imgs(ctx, main_id, transforms: bool, to_8bit: bool, brand: HeifBrand) -> list:
     _result: List[UndecodedHeifFile] = []
-    n = lib.heif_context_get_number_of_top_level_images(ctx)
-    if not n > 1:
+    top_img_count = lib.heif_context_get_number_of_top_level_images(ctx)
+    if not top_img_count > 1:
         return _result
-    top_lvl_image_ids = ffi.new("heif_item_id[]", n)
-    lib.heif_context_get_list_of_top_level_image_IDs(ctx, top_lvl_image_ids, n)
+    top_lvl_image_ids = ffi.new("heif_item_id[]", top_img_count)
+    lib.heif_context_get_list_of_top_level_image_IDs(ctx, top_lvl_image_ids, top_img_count)
     for _image_id in top_lvl_image_ids:
         if _image_id == main_id:
             continue