+++ pillow_heif/__init__.py
@@ -2,8 +2,6 @@
 Import all possible stuff that can be used.
 """
 
-# pylint: disable=unused-import
-# pylint: disable=redefined-builtin
 
 from ._lib_info import (
     have_decoder_for_format,
@@ -14,8 +12,24 @@ from ._lib_info import (
 from ._options import options
 from ._version import __version__
 from .as_opener import HeifImageFile, check_heif_magic, register_heif_opener
-from .constants import *  # pylint: disable=unused-wildcard-import
+from .constants import heif_filetype_maybe  # DEPRECATED
+from .constants import heif_filetype_no  # DEPRECATED
+from .constants import heif_filetype_yes_supported  # DEPRECATED
+from .constants import heif_filetype_yes_unsupported  # DEPRECATED
+from .constants import (
+    HeifBrand,
+    HeifChannel,
+    HeifChroma,
+    HeifColorProfileType,
+    HeifColorspace,
+    HeifCompressionFormat,
+    HeifErrorCode,
+    HeifFiletype,
+)
 from .error import HeifError
+from .misc import reset_orientation
+
+# pylint: disable=redefined-builtin
 from .reader import (
     HeifFile,
     HeifThumbnail,